RFC: Check in compressed and optimized member headshots (#254)
### Summary

Right now development experience on the repo sucks whenever you need to test something related to images. Right now the images are stored in a AWS S3 bucket with full resolution, resized, optimized by imgix which we pay $10 each month. Changing an image involves multiple steps:

1. add to S3;
2. purge cache in imgix;
3. commit, push, and wait for 6-7 minutes until CI builds everything;
4. If you found a problem in step 3, repeat the entire thing and waste another 10 minutes.

I tried git LFS, but it doesn't work well with the CI pipeline and netlify's permission model. Therefore, I believe the only way to achieve a better local dev experience is to check-in **compressed, optimized** assets. It turns out that generating those optimized assets as a much easier job than the 4-step process above. It only has two steps, at least on mac:

1. Select all images, open all of them by preview, mass resize them to height 500.
2. Drag and drop all the images into [imageoptim mac app](https://imageoptim.com/)

The produced images now have the same size and quality as the one produced by imgix. Given our use case, even if we refuse to use mac, and use the imageoptim web API only, it will still be much cheaper: we pay $12 whenever we need it, instead of $10 every month.

All the added images only increased the size by 1.3MB, so it won't blow up the repo.

### Test Plan

To enforce that we actually compress the image, a script is added to automatically transform all images. The same script can also be used to check whether the member photo's size is below 20KB, which is a good indication of whether things are properly compressed.