Eliminate the need of build:grid (#274)
It turns out the reason we need double `build:grid` is that the first build calls `yarn gen` that produces the generated folder for gridsome to consume. We can avoid that by running a cheap `yarn gen` before `yarn build:grid`. Now we can cut the build step time of preview by 1/3.