Bump typescript-eslint to 3.0.1 (#311)
The new `typescript-eslint` changes a few defaults and renames some rules, so some manual linter fixes are necessary.

The most notable changes is that it changes the name of rule that enforces you annotate functions as `@typescript-eslint/explicit-module-boundary-types`. Now they exist as warnings. I think it's a good practice to annotate them anyways, so that readers of the code don't have to redo type inference in their head. Therefore, I annotated as much as I can.