Eliminate strings-domain component usages (#351)
## Summary

The `strings-domain` component also doesn't work well with TS and results in many red squiggly lines. There is currently no way in Vue 2 to make this well typed, because the slots are not part of the type signature of vue components, so there is no way to communicate to the type system that the type of the prop `value` is linked to the type of the named slot `key`.

In this PR, I removed them all and just replaced them with normal field access. As a result, the code should be more readable, and it should be more obvious what kind of data we are using from the json.

## Note

This PR also removes support for linking medium article in the project page, because no project is using it and it probably won't be used later as well.